chore(headless): add generic addPageView instead
