+++ packages/atomic/src/components/ipx/atomic-ipx-result-link/atomic-ipx-result-link.tsx
@@ -1,5 +1,8 @@
 import {isUndefined} from '@coveo/bueno';
-import {loadIPXActionsHistoryActions} from '@coveo/headless';
+import {
+  GenericAnalyticsActionCreators,
+  loadGenericAnalyticsActions,
+} from '@coveo/headless';
 import {Component, h, Prop, Element} from '@stencil/core';
 import {Bindings} from '../../../components';
 import {buildCustomEvent} from '../../../utils/event-utils';
@@ -56,6 +59,7 @@ export class AtomicIPXResultLink implements InitializableComponent {
   private hasDefaultSlot!: boolean;
   private linkAttributes?: Attr[];
   private stopPropagation?: boolean;
+  private actionsHistoryActions?: GenericAnalyticsActionCreators;
 
   public initialize() {
     this.host.dispatchEvent(
@@ -65,6 +69,9 @@ export class AtomicIPXResultLink implements InitializableComponent {
           this.stopPropagation = stopPropagation;
         }
       )
+    );
+    this.actionsHistoryActions = loadGenericAnalyticsActions(
+      this.bindings.engine
     );
   }
 
@@ -76,10 +83,11 @@ export class AtomicIPXResultLink implements InitializableComponent {
 
   public async onSelect() {
     const resultPermanentId = this.result.raw.permanentid;
-    if (resultPermanentId) {
-      const action = loadIPXActionsHistoryActions(
-        this.bindings.engine
-      ).addPageViewEntryInActionsHistory(resultPermanentId);
+    if (resultPermanentId && this.actionsHistoryActions) {
+      const action =
+        this.actionsHistoryActions.addPageViewEntryInActionsHistory(
+          resultPermanentId
+        );
       this.bindings.engine.dispatch(action);
     }
     this.interactiveResult.select();
