+++ packages/headless/src/features/analytics/analytics-actions.ts
@@ -1,6 +1,9 @@
+import {createAsyncThunk} from '@reduxjs/toolkit';
 import {SearchPageEvents as LegacySearchPageEvents} from 'coveo.analytics/dist/definitions/searchPage/searchPageEvents';
+import {historyStore} from '../../api/analytics/coveo-analytics-utils';
 import {SearchAnalyticsProvider} from '../../api/analytics/search-analytics';
 import {Result} from '../../api/search/search/result';
+import {SearchAppState} from '../../state/search-app-state';
 import {
   validatePayload,
   requiredNonEmptyString,
@@ -199,3 +202,17 @@ export const omniboxFromLink = (
       ),
   };
 };
+
+export const addPageViewEntryInActionsHistory = createAsyncThunk(
+  'analytics/actionsHistory/addPageView',
+  async (itemPermanentId: string, {getState}) => {
+    const state = getState() as SearchAppState;
+    if (state.configuration.analytics.enabled) {
+      historyStore.addElement({
+        name: 'PageView',
+        value: itemPermanentId,
+        time: JSON.stringify(new Date()),
+      });
+    }
+  }
+);
